Set picking a server within specified constraints as the default
This allows for options such as ownership, isp, server bandwidth etc. to be included in the server choice
Moreover, it allows for equal probability of choice between the specified countries/cities, even if the number of servers within them are different. This was not possible using the --countries-as-servers and --cities-as-servers arguments, as both made all the available servers in the specified locations able to be selected. Now the choice happens first between the countries/cities, and only then between the servers in the new location, flattening the probabilities